Add removeItem, update clearAll and add display functions and files
