--- src/crud.js
@@ -1,6 +1,0 @@
-const add = () => {
-    
-    
-}
-
-add();
