+++ src/index.js
@@ -1,16 +1,18 @@
 import './styles.css';
 import status from './status.js';
-import './crud.js';
+import display from './display.js';
+import removeItem from './removeItem.js'
+import clearAll from './clearAll.js'
 
 let listItem = [];
 
 class List {
   constructor(description, index) {
-      this.completed = false;
-      this.description = description;
-      this.index = listItem.length +1;
-      this.id = `id${this.index}`
-  }
+    this.completed = false;
+    this.description = description;
+    this.index = listItem.length +1;
+    this.id = `id${this.index}`;
+  };
 }
 
 const add = () => {
@@ -24,32 +26,6 @@ const add = () => {
 add();
 
 
-const display = () => {
-  const list = JSON.parse(localStorage.getItem('list'));
-  const page = document.getElementById('list')
-  list.forEach((e) => {
-    const div = document.createElement('div');
-    const checkbox = document.createElement('input');
-    const removeBtn = document.createElement('a');
-    removeBtn.innerHTML = '<i class="fas fa-trash"></i>';
-    removeBtn.id = e.id;
-    if (e.completed) {
-      div.style.textDecoration = 'underline line-through';
-      div.classList.add('dim');
-    } else {
-      div.style.textDecoration = 'none';
-    }
-    checkbox.setAttribute('type', 'checkbox');
-    checkbox.name = 'checkbox';
-    checkbox.checked = e.completed;
-    div.append(checkbox);
-    div.append(e.description);
-    div.append(removeBtn);
-    div.classList.add('list-group-item');
-    page.append(div);
-  });
-}
-
 display();
 
 const completed = () => {
@@ -57,47 +33,9 @@ const completed = () => {
 }
 completed();
 
-const clearAll = () => {
-  let list = JSON.parse(localStorage.getItem('list'));
-  document.getElementById('clearBtn').addEventListener('click', () =>{
-    list = list.filter((e) => !e.completed);
-    localStorage.setItem('list', JSON.stringify(list))
-    if (list !== null) {
-      location.reload();
-    }
-  })
-}
+
 clearAll();
 
-// listItem = listItem.sort((a, b) => a.index - b.index);
 
-// if (!localStorage.getItem('list')) {
-//   localStorage.setItem('list', JSON.stringify(listItem));
-// }
-// const populate = () => {
-//   let storeList;
-//   if (localStorage.getItem('list')) {
-//     storeList = JSON.parse(localStorage.getItem('list'));
-//   }
-//   const page = document.getElementById('list');
-//   storeList.forEach((element) => {
-//     const div = document.createElement('div');
-//     const checkbox = document.createElement('input');
-//     if (element.completed) {
-//       (div.style.textDecoration = 'underline line-through');
-//     } else {
-//       (div.style.textDecoration = 'none');
-//     }
-//     checkbox.setAttribute('type', 'checkbox');
-//     checkbox.name = 'checkbox';
-//     checkbox.checked = element.completed;
-//     div.append(checkbox);
-//     div.append(element.description);
-//     div.classList.add('list-group-item');
-//     page.append(div);
-//   });
-// };
 
-// populate();
-
-// status();
+removeItem();
